+++ apps/slackbot-api/src/app/core/utils/index.ts
@@ -1,2 +1,2 @@
 export { SlackHelper } from './slack-helper.utils';
-export { addViewedBySnippetToBlock } from './slack-helper.utils';
+export { addViewedBySnippetToBlock } from './add-viewed-by-snippet-to-block.util';
