+++ apps/slackbot-api/src/app/core/utils/slack-helper.utils.ts
@@ -135,24 +135,3 @@ function applyFilter(items: Player[], filter: IDefaultLeaderboardOpts): Player[]
 
   return result;
 }
-
-// TODO: refactor into SlackApi -> Core
-export function addViewedBySnippetToBlock(blocks: any[], slackUserId: string): any[] {
-  if (!slackUserId) {
-    return blocks;
-  }
-  const response = [
-    ...blocks,
-    {
-      type: 'context',
-      elements: [
-        {
-          type: 'mrkdwn',
-          text: `Requested by <@${slackUserId}>`,
-        },
-      ],
-    },
-  ];
-  console.log('addViewedBySnippetToBlock', response);
-  return response;
-}
